+++ src/app/components/catalog-browser/catalog-browser-details/catalog-browser-details.component.ts
@@ -133,7 +133,6 @@ export class CatalogBrowserDetailsComponent {
    * */
   constructOfferNegotiationRequest(
     offer: Offer,
-    distribution: Distribution,
     datasetId: string
   ): any {
     const endpointURL = distribution.accessService.endpointURL;
@@ -175,10 +174,9 @@ export class CatalogBrowserDetailsComponent {
    * Starts the contract negotiation process.
    */
   startContractNegotiation(datasetId: string): void {
-    if (this.selectedOffer && this.selectedDistribution) {
+    if (this.selectedOffer) {
       const negotiationRequest = this.constructOfferNegotiationRequest(
         this.selectedOffer,
-        this.selectedDistribution,
         datasetId
       );
 
@@ -196,7 +194,7 @@ export class CatalogBrowserDetailsComponent {
         });
     } else {
       this.snackBarService.openSnackBar(
-        'Please select both an offer and a distribution.',
+        'Please select an offer.',
         'OK',
         'center',
         'bottom',